Changed build.gradle.kts and moved .class
