+++ Task_1_1_1/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
-    id("java")
+    id 'java'
+    id 'jacoco'
 }
 
 group = "org.example"
@@ -16,4 +17,10 @@ dependencies {
 
 tasks.test {
     useJUnitPlatform()
+}
+
+jacocoTestReport {
+    reports {
+        xml.required = true
+    }
 }
